UI for Oregon Trail-inspired game
The UI this program creates assembles a team of five. It allows the user
to purchase various types of food and clothing, each of which will have
a unique effect on a team member's health and experience level. The
program also allows the team to choose from a list of cities to travel
to, then updates the players health and the expiration dates of the food
to reflect how much time has passed during their travels.